httpd: rediff patch for blfs layout again
Still remove the ownership changes when installing files: this is
not the current practice, and not needed if installing as root
anyway. And it prevents doing a DESTDIR install as user.
But the build system uses "cp -p" for installing, which keeps
the original user ownership, so remove the -p flag.