coreutils: Revise i18n patch
I've rebased the i18n patch and created coreutils-9.5-i18n-1.patch but
there are two problems in the mbfile.h gnulib header shipped in the
patch.  One is I used unsound hack to avoid introducing mbfile.c but
GCC 14 warns about it with a good reason.  Another is more serious:
the mbfile.h is derived from an old revision of gnulib and it FTBFS on
32-bit x86 with GCC 14.

Sync the patch from Fedora to resolve the issues.